requirements.txt
flask
flask_wtf
flask_sqlalchemy
flask_login
flask_session
email_validator
environs